[feat] 선택 작업 버튼 영역에 Collapse 컴포넌트 적용
