+++ package.json
@@ -9,10 +9,13 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
     "@fortawesome/fontawesome-svg-core": "^6.7.2",
     "@fortawesome/free-solid-svg-icons": "^6.7.2",
     "@fortawesome/react-fontawesome": "^0.2.2",
     "@midudev/tailwind-animations": "^0.1.0",
+    "@mui/material": "^6.4.2",
     "axios": "^1.7.9",
     "lucide-react": "^0.473.0",
     "next": "15.1.4",
