+++ src/utils/currencySymbols.ts
@@ -1,0 +1,13 @@
+export const currencySymbols = {
+    USD: "$",
+    KRW: "₩",
+    EUR: "€",
+    GBP: "£",
+    JPY: "¥",
+    CAD: "C$",
+    AUD: "A$",
+    CNY: "¥",
+    CHF: "CHF",
+    INR: "₹",
+    SGD: "S$",
+};
